finally rental plan works

+++ app/src/main/java/com/rentmycar/rentmycar/domain/model/RentalPlan.kt
@@ -6,7 +6,7 @@ data class RentalPlan(
     val id: Int?,
     val availableFrom: String,
     val availableUntil: String,
-    val price: BigDecimal,
+    val price: Double,
     val userId: Int,
     val carId: Int,
     val createdAt: String,
