+++ app/src/main/java/com/rentmycar/rentmycar/network/response/GetRentalPlanResponse.kt
@@ -1,12 +1,10 @@
 package com.rentmycar.rentmycar.network.response
 
-import java.math.BigDecimal
-
 data class GetRentalPlanResponse(
     val id: Int,
     val availableFrom: String,
     val availableUntil: String,
-    val price: BigDecimal,
+    val price: Double,
     val userId: Int,
     val carId: Int,
     val createdAt: String,
